Move the property initialization onto the traits
