+++ src/main/groovy/wooga/gradle/github/base/tasks/internal/AbstractGithubTask.groovy
@@ -32,112 +32,10 @@ import wooga.gradle.github.base.GithubSpec
 
 abstract class AbstractGithubTask<T extends AbstractGithubTask> extends ConventionTask implements GithubSpec {
 
-    private final Property<String> repositoryName
-
-    @Input
-    @Override
-    Property<String> getRepositoryName() {
-        repositoryName
-    }
-
-    @Override
-    void setRepositoryName(Provider<String> value) {
-        repositoryName.set(value)
-    }
-
-    void setRepositoryName(String name) {
-        if (!GithubRepositoryValidator.validateRepositoryName(name)) {
-            throw new IllegalArgumentException("Repository value '$name' is not a valid github repository name. Expecting `owner/repo`.")
-        }
-
-        repositoryName.set(name)
-    }
-
-    @Optional
-    private final Property<String> baseUrl
-
-    @Input
-    @Override
-    Property<String> getBaseUrl() {
-        baseUrl
-    }
-
-    @Override
-    void setBaseUrl(Provider<String> value) {
-        baseUrl.set(value)
-    }
-
-    @Optional
-    private final Property<String> username
-
-    @Input
-    @Override
-    Property<String> getUsername() {
-        username
-    }
-
-    @Override
-    void setUsername(Provider<String> value) {
-        username.set(value)
-    }
-
-    @Optional
-    private final Property<String> password
-
-    @Input
-    @Override
-    Property<String> getPassword() {
-        password
-    }
-
-    @Override
-    void setPassword(Provider<String> value) {
-        password.set(value)
-    }
-
-    @Optional
-    private final Property<String> token
-
-    @Input
-    @Override
-    Property<String> getToken() {
-        token
-    }
-
-    @Override
-    void setToken(Provider<String> value) {
-        token.set(value)
-    }
-
-    protected final Property<GitHub> clientProvider
-
-    @Internal
-    @Override
-    Property<GitHub> getClientProvider() {
-        clientProvider
-    }
-
-    protected final Property<String> branchName
-
-    @Internal
-    @Override
-    Property<String> getBranchName() {
-        return branchName
-    }
-
     private final Class<T> taskType
 
     AbstractGithubTask(Class<T> taskType) {
         this.taskType = taskType
-
-        repositoryName = project.objects.property(String)
-        baseUrl = project.objects.property(String)
-        username = project.objects.property(String)
-        password = project.objects.property(String)
-        token = project.objects.property(String)
-        clientProvider = project.objects.property(GitHub)
-        clientProvider.set(GithubClientFactory.clientProvider(username, password, token))
-        branchName = project.objects.property(String)
     }
 
     @Internal
